hsmodule: move auto-startup query to settings parsing logic
This is to prepare the stage for removing the pre-baked settings.ini
file altogether. This check is too early, before there is an attempt
to read (and therefore generate) a valid settings.ini so it'll try to
substitute a pair in an empty file...

Let LES write default values, then prompt the user.